+++ setup.cfg
@@ -47,4 +47,4 @@ ignore =
     W503
 extend-exclude = tests/*.py,.venv/*.py,.tox/*.py
 filename = ./src/*.py
-max-line-length = 100
+max-line-length = 105
