+++ src/nerdle/scoring.py
@@ -1,11 +1,9 @@
-from collections import defaultdict
 from functools import lru_cache
-from typing import Dict, Set
 
 import numpy as np
 from numba import int8, int32, jit
 
-from .words import Word, WordSeries
+from .words import Word
 
 
 class Scorer:
@@ -25,34 +23,6 @@ class Scorer:
 
     def score_word(self, solution: Word, guess: Word) -> int:
         return score_word_jit(solution.vector, guess.vector, self._powers)
-
-    def get_solutions_by_score(self, potential_solns: WordSeries, guess: Word) -> Dict[int, WordSeries]:
-
-        score_func = np.vectorize(self.score_word)
-        scores = score_func(potential_solns.words, guess)
-        unique_scores, positions = np.unique(scores, return_inverse=True)
-
-        solns_by_score: dict[int, WordSeries] = {}
-        for (i, unique_score) in enumerate(unique_scores):
-            maps_to_score = positions == i
-            filtered_solns = potential_solns[maps_to_score]
-            solns_by_score[unique_score] = filtered_solns
-
-        return solns_by_score
-
-    def get_solutions_by_score_old(self, potential_solns: Set[str], guess: str) -> Dict[int, Set[str]]:
-        solns_by_score = defaultdict(set)
-        for soln in potential_solns:
-            score = self.score_word(soln, guess)
-            solns_by_score[score].add(soln)
-
-        return solns_by_score
-
-    def get_histogram(self, potential_solns: WordSeries, guess: Word) -> Dict[int, int]:
-        score_func = np.vectorize(self.score_word)
-        scores = score_func(potential_solns.words, guess)
-        unique_scores, counts = np.unique(scores, return_counts=True)
-        return {score: count for score, count in zip(unique_scores, counts)}
 
 
 def get_fast_ternary_lookup(size: int) -> np.ndarray:
